fix: Update vendor dashboard template loader and improve field styling. (#3006)
* fix: update vendor dashboard template loader and improve select2 field styling

* fix: update vendor dashboard template loader and improve select2 field styling

* fix: update namespace reference for FullWidthVendorLayout usage in CommonServiceProvider

* fix: add missing margin reset to Sidebar navigation list styles

* fix: add missing comma to service provider, refactor and update hooks in FullWidthVendorLayout

* fix: adjust mobile padding for vendor dashboard section headers
