+++ assets/src/less/dashboard.less
@@ -539,6 +539,10 @@ div.media-sidebar a.edit-attachment{
             width: calc(100% - var(--dokan-sidebar-width, 250px));
         }
     }
+
+    .select2-search__field {
+        min-width: 220px !important;
+    }
 }
 
 .dokan-column-name-with-avatar {
