+++ includes/DependencyManagement/Providers/CommonServiceProvider.php
@@ -24,6 +24,7 @@ class CommonServiceProvider extends BaseServiceProvider {
         \WeDevs\Dokan\Commission\RecalculateCommissions::class,
         \WeDevs\Dokan\Order\RefundHandler::class,
         \WeDevs\Dokan\Exceptions\Handler::class,
+        \WeDevs\Dokan\Shortcodes\FullWidthVendorLayout::class,
 	];
 
 	/**
